Got the custom notifications working!
Made the custom notifications work, both in the settings view as well
as what is displayed in the notification alert body.

+++ Sunset/Sunset/ViewController.m
@@ -98,7 +98,7 @@
 }
 
 - (void)setNotifications {
-  [sunEventObject setNotificationsWithSeconds: 60 * [[myDefaults objectForKey:@"notificationTimeCustomization"] integerValue]];
+  [sunEventObject setNotificationsWithSeconds: (int) (60 * [[myDefaults objectForKey:@"notificationTimeCustomization"] integerValue])];
 }
 
 - (void)viewDidLoad {
